Provide default Rails configuration
The application has been provided with a default configuration for Rails. Localization has been enabled, and the default locale has been set to English. The database defaults to Postgresql in production, and expects its configuration to come from environment variables. Last but not least, the production environment has also been configured to require SSL. Since the application stores passwords, it is absolutely necessary to secure all communication between a client and the server with encryption.