Install Bourbon family as the project's CSS library
This project makes use of the Bourbon family as its CSS library. The Bourbon family comes with a lot of really neat features that make it easy to design and build nice looking web pages. All gems of the family are all written in SCSS, which makes customization really easy.