Ignore version files when deploying the application
Capistrano can ignore files when deploying the application to a server. This is useful to exclude the spec directory, for example, and it its particularly useful for ignoring the .ruby-version and .ruby-gemset files. While these files are very useful for development purposes, they can be problematic on production machines where different Ruby versions are used. Passenger for example picks up the .ruby-version file, and uses it to determine which version of Ruby to run. When the Ruby version on the server doesn't match the version specified in the file, Passenger fails to load the application because it looks for the gems in the wrong path.

By removing the reference to a specific Ruby version, we allow users to deploy the application to any server running Ruby 2.x.x.