Change color of delete buttons
To decrease the risk of users accidentally clicking buttons that delete
resources, the color of those buttons has been changed. To enable the
easy creation of other button types as well, the styling of buttons has
been moved to a mixin that takes a color as its argument. This way,
totally different buttons can be made by simple including the mixin with
the new button types color.