Fix broken XHA form
The form to create XHA interfaces was broken. The select_tag method
created the right dropdown menu, but it did not set the ID correctly, so
that the controller pick up the selection from the params hash
automatically.

The mistake was that select_tag was used instead of f.select, which
creates the same dropdown menu, but nests it inside the xha[...]
namespace.

By changing the select_tag method to f.select, the bug is fixed and the
form works correctly again. To detect this error in future releases, the
view specs have been adjusted as well to verify that the selection is
nested correctly.