Save associations for servers
When creating or editing a server resource, its associations are not
saved. This makes it impossible to assign a status, model or location to
any server.

This bug is caused by an error in the servers' form: The selection menu
is built using the 'select_tag' method, and not with 'f.select'. Only
the latter wraps the selection menu inside the :adonis or :proteus hash,
which is expected by the controller. Due to this bug, the selection was
never processed.

To fix the bug, the form has been changed to use 'f.select'.
Furthermore, specs have been written to verify that the associations are
saved correctly when creating or editing servers. The view specs have
been adjusted as well work with the selections' changed identifiers.

Fixes gh-2